Auxiliary for controlling IB

Auxiliary function for ib bias correction.

ibControl(
  tol = 1e-05,
  maxit = 25,
  verbose = FALSE,
  seed = 123L,
  H = 1L,
  constraint = TRUE,
  early_stop = FALSE,
  cens = FALSE,
  right = NULL,
  left = NULL,
  mis = FALSE,
  prop = NULL,
  out = FALSE,
  eps = NULL,
  G = NULL,
  func = function(x) rowMeans(x, na.rm = T),
  sim = NULL
)

Arguments

tol

positive convergence tolerance \(\epsilon\). The ib procedure converges when \(||\hat{\theta}^{k+1}-\hat{\theta}^k||_2/p<\epsilon\), where \(p\) is the dimension of \(\theta\).

maxit

integer representing the maximal number of iterations.

verbose

if TRUE, it prints some output in the console at each iteration.

seed

integer to set the seed (see Random).

H

integer representing the number of bootstrap estimates (see ib).

constraint

if TRUE (default), constraint for extra_param is used in the iterative procedure (see 'Details' of ib).

early_stop

if TRUE (default is FALSE), the iterative procedure stops as soon as there is no improvment in the minimization of the objective function (see 'Details' of ib).

cens

if TRUE the simulated responses are censored according to left and right values.

right

double for right-censoring (only used if cens=TRUE).

left

double for left-censoring (only used if cens=TRUE).

mis

if TRUE the simulated responses have missing data at random.

prop

double between 0 and 1 representing the proportion of missing data (only used if mis=TRUE).

out

if TRUE the simulated responses are also generated with a contamination mechanism.

eps

double between 0 and 1 representing the proportion of outliers in the data (only used if out=TRUE).

G

a function to generate outliers. It takes only a sample size as argument.

func

a function to reduce the H bootstrap estimates (rowwise). By default, the average is computed. The user can supply a function. One could imagine using other function such as the median or a trimmed mean.

sim

a user-defined function for simulating responses (see 'Details')

Value

a list with components named as the arguments.

Details

sim allows the user to provide its own function for generating responses. Currently it is only supported for generalized linear models with the prototype `fun(object, control, extra_param, ...)` (see ib).

See also